requirements.txt
dbt-core
dbt-postgres
dbt-snowflake
python-dotenv